+++ openvair/modules/volume/domain/physical_fs/localfs.py
@@ -9,9 +9,12 @@ Classes:
 """
 
 from typing import Any, Dict
+from pathlib import Path
 
 from openvair.libs.log import get_logger
-from openvair.modules.tools.utils import execute
+from openvair.libs.cli.models import ExecuteParams
+from openvair.libs.cli.executor import execute
+from openvair.libs.cli.exceptions import ExecuteError
 from openvair.modules.volume.domain.base import BaseVolume
 
 LOG = get_logger(__name__)
@@ -88,18 +91,27 @@ class LocalFSVolume(BaseLocalFSVolume):
             f'size={self.size}, format={self.format}'
         )
         preallocation = self.provisioning if self.format != 'raw' else 'off'
-        volume_path = f'{self.path}/volume-{self.id}'
-        execute(
-            'qemu-img',
-            'create',
-            '-f',
-            self.format,
-            '-o',
-            f'preallocation={preallocation}',
-            volume_path,
-            self.size,
-            run_as_root=self._execute_as_root,
-        )
+        volume_path = Path(self.path, f'volume-{self.id}')
+        try:
+            execute(
+                'qemu-img',
+                'create',
+                '-f',
+                self.format,
+                '-o',
+                f'preallocation={preallocation}',
+                str(volume_path),
+                self.size,
+                params=ExecuteParams(  # noqa: S604
+                    run_as_root=self._execute_as_root,
+                    shell=True,
+                    raise_on_error=True,
+                )
+            )
+        except (ExecuteError, OSError) as err:
+            LOG.error(f'Error while creating volume: {err!s}')
+            raise
+
         return self.__dict__
 
     def delete(self) -> Dict:
@@ -109,8 +121,22 @@ class LocalFSVolume(BaseLocalFSVolume):
             Dict: A dictionary representation of the volume's attributes.
         """
         LOG.info(f'Deleting volume with id={self.id}, path={self.path}')
-        volume_path = f'{self.path}/volume-{self.id}'
-        execute('rm', '-f', volume_path, run_as_root=self._execute_as_root)
+        volume_path = Path(self.path, f'volume-{self.id}')
+        try:
+            execute(
+                'rm',
+                '-f',
+                str(volume_path),
+                params=ExecuteParams(  # noqa: S604
+                    run_as_root=self._execute_as_root,
+                    shell=True,
+                    raise_on_error=True,
+                )
+            )
+        except (ExecuteError, OSError) as err:
+            LOG.error(f'Error while deleting volume: {err!s}')
+            raise
+
         return self.__dict__
 
     def extend(self, new_size: str) -> Dict:
@@ -127,14 +153,23 @@ class LocalFSVolume(BaseLocalFSVolume):
             f'path={self.path} to new size={new_size}'
         )
         self._check_volume_exists()
-        volume_path = f'{self.path}/volume-{self.id}'
-        execute(
-            'qemu-img',
-            'resize',
-            volume_path,
-            new_size,
-            run_as_root=self._execute_as_root,
-        )
+        volume_path = Path(self.path, f'volume-{self.id}')
+        try:
+            execute(
+                'qemu-img',
+                'resize',
+                str(volume_path),
+                new_size,
+                params=ExecuteParams(  # noqa: S604
+                    run_as_root=self._execute_as_root,
+                    shell=True,
+                    raise_on_error=True,
+                )
+            )
+        except (ExecuteError, OSError) as err:
+            LOG.error(f'Error while extending volume: {err!s}')
+            raise
+
         return self.__dict__
 
     def attach_volume_info(self) -> Dict:
