+++ openvair/modules/volume/domain/remotefs/nfs.py
@@ -8,9 +8,12 @@ Classes:
 """
 
 from typing import Any, Dict
+from pathlib import Path
 
 from openvair.libs.log import get_logger
-from openvair.modules.tools.utils import execute
+from openvair.libs.cli.models import ExecuteParams
+from openvair.libs.cli.executor import execute
+from openvair.libs.cli.exceptions import ExecuteError
 from openvair.modules.volume.domain.base import RemoteFSVolume
 from openvair.modules.volume.domain.remotefs import exceptions
 
@@ -47,22 +50,28 @@ class NfsVolume(RemoteFSVolume):
             Dict: A dictionary representation of the volume's attributes.
         """
         preallocation = self.provisioning if self.format != 'raw' else 'off'
-        volume_path = f'{self.path}/volume-{self.id}'
-        _, err = execute(
-            'qemu-img',
-            'create',
-            '-f',
-            self.format,
-            '-o',
-            f'preallocation={preallocation}',
-            volume_path,
-            self.size,
-            run_as_root=self._execute_as_root,
-        )
-        if err:
+        volume_path = Path(self.path, f'volume-{self.id}')
+        try:
+            execute(
+                'qemu-img',
+                'create',
+                '-f',
+                self.format,
+                '-o',
+                f'preallocation={preallocation}',
+                str(volume_path),
+                self.size,
+                params=ExecuteParams(  # noqa: S604
+                    run_as_root=self._execute_as_root,
+                    shell=True,
+                    raise_on_error=True,
+                )
+            )
+        except (ExecuteError, OSError) as err:
             message = f'Qemu img create raised exception: {err!s}'
             LOG.error(message)
             raise exceptions.QemuImgCreateException(message)
+
         LOG.info(f'Created volume {volume_path}')
         return self.__dict__
 
@@ -76,17 +85,23 @@ class NfsVolume(RemoteFSVolume):
         Returns:
             Dict: A dictionary representation of the volume's attributes.
         """
-        volume_path = f'{self.path}/volume-{self.id}'
-        _, err = execute(
-            'rm',
-            '-f',
-            volume_path,
-            run_as_root=self._execute_as_root,
-        )
-        if err:
+        volume_path = Path(self.path, f'volume-{self.id}')
+        try:
+            execute(
+                'rm',
+                '-f',
+                str(volume_path),
+                params=ExecuteParams(  # noqa: S604
+                    run_as_root=self._execute_as_root,
+                    shell=True,
+                    raise_on_error=True,
+                )
+            )
+        except (ExecuteError, OSError) as err:
             message = f'Delete volume raised exception: {err!s}'
             LOG.error(message)
             raise exceptions.DeleteVolumeException(message)
+
         LOG.info(f'Deleted volume {volume_path}')
         return self.__dict__
 
@@ -104,18 +119,20 @@ class NfsVolume(RemoteFSVolume):
             Dict: A dictionary representation of the volume's attributes.
         """
         self._check_volume_exists()
-        volume_path = f'{self.path}/volume-{self.id}'
-        _, err = execute(
-            'qemu-img',
-            'resize',
-            volume_path,
-            new_size,
-            run_as_root=self._execute_as_root,
-        )
-        if err:
+        volume_path = Path(self.path, f'volume-{self.id}')
+        try:
+            execute(
+                'qemu-img',
+                'resize',
+                str(volume_path),
+                new_size,
+                run_as_root=self._execute_as_root,
+            )
+        except (ExecuteError, OSError) as err:
             message = f'Qemu img extend volume raised exception: {err!s}'
             LOG.error(message)
             raise exceptions.QemuImgExtendException(message)
+
         LOG.info(f'Extended volume {volume_path} to size {new_size}')
         return self.__dict__
